web_demo_cocktails/requirements.txt
Flask==2.1.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.0.1
gunicorn==20.1.0
numpy==1.19.5
openvino==2022.1.0
Pillow==9.1.1
WTForms==3.0.1